Move add_controlled_RX_gate to quri-parts-circuit.utils
